Impl SimpleHasher for Digest
