+++ Cargo.toml
@@ -60,6 +60,7 @@ rocksdb = { version = "0.19.0", features = ["lz4"] }
 serde = { version = "1.0.137", features = ["derive", "rc"] }
 serde_json = { version = "1.0" }
 sha2 = "0.10.6"
+digest = "0.10.6"
 thiserror = "1.0.38"
 tiny-keccak = "2.0.2"
 tracing = "0.1.37"
