+++ rollup-interface/src/state_machine/crypto/mod.rs
@@ -1,0 +1,2 @@
+mod simple_hasher;
+pub use simple_hasher::{NoOpHasher, SimpleHasher};
