+++ rollup-interface/src/state_machine/zk/traits.rs
@@ -3,6 +3,8 @@ use core::fmt::Debug;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 
+use crate::crypto::SimpleHasher;
+
 /// A trait implemented by the prover ("host") of a zkVM program.
 pub trait ZkvmHost: Zkvm {
     /// Give the guest a piece of advice non-deterministically
@@ -35,52 +37,6 @@ pub trait ZkvmGuest: Zkvm {
     fn commit<T: Serialize>(&self, item: &T);
 }
 
-/// A minimal trait representing a hash function. We implement our own
-/// rather than relying on `Digest` for broader compatibility.
-pub trait SimpleHasher: Sized {
-    /// Creates a new hasher with default state.
-    fn new() -> Self;
-    /// Ingests the provided data, updating the hasher's state.
-    fn update(&mut self, data: &[u8]);
-    /// Consumes the hasher state to produce a digest.
-    fn finalize(self) -> [u8; 32];
-    /// Returns the digest of the provided data.
-    fn hash(data: impl AsRef<[u8]>) -> [u8; 32] {
-        let mut hasher = Self::new();
-        hasher.update(data.as_ref());
-        hasher.finalize()
-    }
-}
-
-pub struct NoOpHasher;
-impl SimpleHasher for NoOpHasher {
-    fn new() -> Self {
-        Self
-    }
-
-    fn update(&mut self, _data: &[u8]) {}
-
-    fn finalize(self) -> [u8; 32] {
-        [0u8; 32]
-    }
-}
-
-// impl<T: Digest> SimpleHasher for T
-// where
-//     [u8; 32]: From<GenericArray<u8, <T as OutputSizeUser>::OutputSize>>,
-// {
-//     fn new() -> Self {
-//         <T as Digest>::new()
-//     }
-
-//     fn update(&mut self, data: &[u8]) {
-//         self.update(data)
-//     }
-
-//     fn finalize(self) -> [u8; 32] {
-//         self.finalize().into()
-//     }
-// }
 /// This trait is implemented on the struct/enum which expresses the validity condition
 pub trait ValidityCondition: Serialize + DeserializeOwned {
     /// Combine two conditions into one (typically run inside a recursive proof).
